Time: 1 ms (100.00%) | Memory: 62.2 MB (51.08%) - LeetSync

+++ leetCodeDaily/4005-maximum-total-subarray-value-i/maximum-total-subarray-value-i.java
@@ -1,7 +1,7 @@
 class Solution {
     public long maxTotalValue(int[] nums, int k) {
-        long max = Integer.MIN_VALUE;
-        long min = Integer.MAX_VALUE;
+        long max = Long.MIN_VALUE;
+        long min = Long.MAX_VALUE;
 
         for(int num:nums){
             if(num > max) max = num;
